versionary: add get_applicable method
which adds the possibility to select a version based on passed
minium and maximum versions requirements,
so we do not need to know what versions are actually implemented and
can select the highest available version within the given boundaries.
The earliest implementation can be picked by choosing func=min.

If no version was found a new exception NoApplicableVersion is raised

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>